refactor: changed token payload structure

+++ src/posts/posts.service.ts
@@ -1,5 +1,6 @@
 import {
   BadRequestException,
+  ForbiddenException,
   Injectable,
   NotFoundException
 } from '@nestjs/common'
@@ -11,15 +12,13 @@ import { Repository } from 'typeorm'
 import { UsersService } from 'src/users/users.service'
 import { Pagination } from 'src/decorators/pagination-params.decorator'
 import { PaginatedResource } from 'src/common/dto/paginated-resourse.dto'
+import { omit } from 'lodash'
 
 @Injectable()
 export class PostsService {
-  constructor(
-    @InjectRepository(Post) private postRepository: Repository<Post>,
-    private userService: UsersService
-  ) {}
+  constructor(@InjectRepository(Post) private postRepository: Repository<Post>, private userService: UsersService) { }
 
-  async create(createPostDto: CreatePostDto, username: string) {
+  async create(createPostDto: CreatePostDto, username: string): Promise<Post> {
     const user = await this.userService.findOne(username)
 
     if (!user) {
@@ -27,7 +26,6 @@ export class PostsService {
     }
 
     const post = this.postRepository.create({ ...createPostDto, user })
-
     const createdPost = await this.postRepository.save(post)
 
     delete createdPost.user
@@ -35,14 +33,13 @@ export class PostsService {
     return createdPost
   }
 
-  async findUserPosts(
-    pagination: Pagination,
-    username: string
-  ): Promise<PaginatedResource<Post>> {
+  async findUserPosts(pagination: Pagination, username: string): Promise<PaginatedResource<Post>> {
     const [posts, total] = await this.postRepository.findAndCount({
       take: pagination.limit,
       skip: pagination.offset,
-      where: { user: { username } }
+      where: {
+        user: { username }
+      }
     })
 
     return {
@@ -53,24 +50,37 @@ export class PostsService {
     }
   }
 
-  findOne(slug: string) {
+  findOne(slug: string): Promise<Post> {
     return this.postRepository.findOne({ where: { slug } })
   }
 
-  async update(slug: string, updatePostDto: UpdatePostDto) {
+  async update(userId: number, slug: string, updatePostDto: UpdatePostDto): Promise<Post> {
+
+    const post = await this.postRepository.findOne({ where: { slug }, relations: ['user'] })
+
+    if (!userId || post.user.id !== userId) {
+      throw new ForbiddenException('У вас нет прав на выполнение этой операции')
+    }
+
     await this.postRepository.update({ slug }, updatePostDto)
     const updatedPost = await this.findOne(slug)
 
     return updatedPost
   }
 
-  async remove(slug: string) {
-    const post = await this.findOne(slug)
+  async remove(slug: string, userId: number): Promise<Omit<Post, 'user'>> {
+    const post = await this.postRepository.findOne({ where: { slug }, relations: ['user'] })
 
     if (!post) {
-      throw new NotFoundException('Пост с таким идентификатором не найден')
+      throw new NotFoundException('Пост не найден')
     }
 
-    return await this.postRepository.remove(post)
+    if (post.user.id !== userId) {
+      throw new ForbiddenException('У вас нет прав на выполнение этой операции')
+    }
+
+    const removedPost = await this.postRepository.remove(post)
+
+    return omit(removedPost, ['user'])
   }
 }
